requirements.txt
PyGithub
PyYAML
yamllint
python-dotenv
pycowsay

